delete unused code and refactor
