+++ pages/_app.js
@@ -4,56 +4,8 @@ import { uid } from "uid";
 import StyledButton from "@/components/button";
 import StyledFooter from "@/components/footer";
 import Link from "next/link";
-
-const dataPrototype = {
-  "9b51189fa12": {
-    id: "9b51189fa12",
-    name: "Einkauf",
-    color: "#c0d313",
-  },
-  b51189fa126: {
-    id: "b51189fa126",
-    name: "Miete",
-    color: "#001DFF",
-  },
-  "51189fa126b": {
-    id: "51189fa126b",
-    name: "Spaß",
-    color: "#FF0C00",
-  },
-  "1189fa126b6": {
-    id: "1189fa126b6",
-    name: "Sonstiges",
-    color: "#00FF19",
-  },
-  be90e393b31: {
-    id: "be90e393b31",
-    name: "Sparen",
-    color: "#FA00FF",
-  },
-};
-const startingInput = {
-  "9b51189fa12": {
-    value: 0,
-    valueIst: 0,
-  },
-  b51189fa126: {
-    value: 0,
-    valueIst: 0,
-  },
-  "51189fa126b": {
-    value: 0,
-    valueIst: 0,
-  },
-  "1189fa126b6": {
-    value: 0,
-    valueIst: 0,
-  },
-  be90e393b31: {
-    value: 0,
-    valueIst: 0,
-  },
-};
+import startingInput from "@/utils/data/starting-input";
+import dataPrototype from "@/utils/data/dataPrototype";
 
 export default function App({ Component, pageProps }) {
   const [data, setData] = useState({});
