+++ utils/data/dataPrototype.js
@@ -1,0 +1,28 @@
+const dataPrototype = {
+  "9b51189fa12": {
+    id: "9b51189fa12",
+    name: "Einkauf",
+    color: "#c0d313",
+  },
+  b51189fa126: {
+    id: "b51189fa126",
+    name: "Miete",
+    color: "#001DFF",
+  },
+  "51189fa126b": {
+    id: "51189fa126b",
+    name: "Spaß",
+    color: "#FF0C00",
+  },
+  "1189fa126b6": {
+    id: "1189fa126b6",
+    name: "Sonstiges",
+    color: "#00FF19",
+  },
+  be90e393b31: {
+    id: "be90e393b31",
+    name: "Sparen",
+    color: "#FA00FF",
+  },
+};
+export default dataPrototype;
